fix(nodejs): compatible with Windows builds
